+++ lib/v2/indienova/article.js
@@ -1,6 +1,6 @@
 const got = require('@/utils/got');
 const cheerio = require('cheerio');
-const url = require('url');
+const { parseList, parseItem } = require('./utils');
 
 module.exports = async (ctx) => {
     const type = ctx.params.type;
@@ -10,43 +10,14 @@ module.exports = async (ctx) => {
     }
     const response = await got(link);
 
-    const host = 'https://indienova.com';
-
     const $ = cheerio.load(response.data);
-    const list = $('.article-panel h4').get();
+    const list = parseList($);
 
-    const items = await Promise.all(
-        list.map(async (item) => {
-            const $$ = cheerio.load(item);
-            const itemUrl = url.resolve(host, $$('a').attr('href'));
-            const title = $$('a').text();
-
-            const cache = await ctx.cache.get(itemUrl);
-            if (cache) {
-                return JSON.parse(cache);
-            }
-
-            const resp = await got.get(itemUrl);
-
-            const $$$ = cheerio.load(resp.data);
-
-            const description = $$$('.indienova-single-post').html();
-
-            const single = {
-                title,
-                description,
-                pubDate: '',
-                link: itemUrl,
-                author: '',
-            };
-            ctx.cache.set(itemUrl, JSON.stringify(single));
-            return single;
-        })
-    );
+    const items = await Promise.all(list.map((item) => ctx.cache.tryGet(item.link, () => parseItem(item))));
 
     ctx.state.data = {
-        title: 'INDIENOVA',
-        link: 'https://www.indienova.com/indie-game-news/',
+        title: $('head title').text(),
+        link,
         description: '独立游戏资讯 | indienova 独立游戏',
         item: items,
     };
