+++ lib/v2/indienova/maintainer.js
@@ -1,0 +1,6 @@
+module.exports = {
+    '/column/:columnId': ['TonyRL'],
+    '/gamedb/recent': ['TonyRL'],
+    '/usergames': ['TonyRL'],
+    '/:type': ['GensouSakuya', 'kt286'],
+};
