+++ lib/v2/indienova/router.js
@@ -2,4 +2,5 @@ module.exports = (router) => {
     router.get('/column/:columnId', require('./column'));
     router.get('/gamedb/recent', require('./gamedb'));
     router.get('/usergames', require('./usergames'));
+    router.get('/:type', require('./article'));
 };
