+++ website/docs/routes/game.mdx
@@ -273,7 +273,7 @@
 
 ### 文章 {#indienova-du-li-you-xi-wen-zhang}
 
-<Route author="GensouSakuya kt286" example="/indienova/article" path="indienova/:type" paramsDesc={['类型: `article` 文章，`development` 开发']} />
+<Route author="GensouSakuya kt286" example="/indienova/article" path="indienova/:type" paramsDesc={['类型: `article` 文章，`development` 开发']} radar="1" />
 
 ### 专题 {#indienova-du-li-you-xi-zhuan-ti}
 
@@ -321,7 +321,7 @@
 
 ### GameDB 游戏库 {#indienova-du-li-you-xi-gamedb-you-xi-ku}
 
-<Route author="TonyRL" example="/indienova/gamedb/recent" path="/indienova/gamedb/recent/:platform?" paramsDesc={['平台，留空']} radar="1">
+<Route author="TonyRL" example="/indienova/gamedb/recent" path="/indienova/gamedb/recent/:platform?" paramsDesc={['平台，见下表，留空为 `all`']} radar="1">
   | 全部 | PS4 | XBOX One | Nintendo Switch |
   | ---- | --- | -------- | --------------- |
   | all  | ps4 | xboxone  | nintendo-switch |
